fix: takeout separated index service of order module
